Выносит логику bigPicture в отдельный модуль
